Changed the definition of the comparison theorem for paths.
Before the change, the comparison were large. The theorem stated that
for two paths, either one was a large prefix of the other, or both were
disjoint.

Now by default, comparisons are strict. For two paths, either they are equal,
or one is a strict prefix, or both are disjoint.