Changed the evaluation of dereferencement of a pointer.
When a location is find as spath q, instead of returning q, returns
q +++ [0] (the spath of the value inside the loc). Changed the proofs
accordingly.

Also, changed the statement of the theorem
`eval_place_mut_borrow_to_ptr` so that it's proven for arbitrary
permission (and not just `Mut`).