Removed a dangerous hypothesis.
The state ?Sm is to be determined. We should not access it. The
hypothesis (fresh_anon Sm a) is at risk of being used by the tactic
eauto, and the state Sm to be defined as a wrong state.

Plus, we should not *use* the tactic (fresh_anon Sm a), this is
something we should derive when we know the state Sm.