Removed a too strong hypothesis.
When applying the rule Leq_FreshMutLoan, the value declared as borrowed
needed to contain no bottoms. I think this hypothesis is not incorrect,
but for that I should have modified the rules to end borrows (to ensure
that it does not contain loans). As this does not seem to be necessary,
I'm dropping the hypothesis and replacing it with a weaker validity
condition.